requirements.txt
boto3
openai
python-dotenv
requests
pynacl